Introducing postpro package and adding fatigue (#48)
